Driver's common attributes-methods grouped into a base class
