+++ src/pymodaq_plugins_keithley/hardware/keithley27XX/keithley27XX_VISADriver.py
@@ -5,19 +5,9 @@ from pymodaq.utils.logger import set_logger, get_module_name
 logger = set_logger(get_module_name(__file__))
 
 
-class Keithley2700VISADriver:
-    """VISA class driver for the Keithley 2700 Multimeter/Switch System
-
-    This class relies on pyvisa module to communicate with the instrument via VISA protocol.
-    Please refer to the instrument reference manual available at:
-    https://download.tek.com/manual/2700-900-01K_Feb_2016.pdf
+class Keithley27XXVISADriver:
+    """VISA base class driver for Keithley 2700-2701-2750 drivers
     """
-    # List the Keithley instruments the user has configured from the .toml configuration file
-    list_instruments = {}
-    for instr in config["Keithley", "2700"].keys():
-        if "INSTRUMENT" in instr:
-            list_instruments[instr] = config["Keithley", "2700", instr, "rsrc_name"]
-    logger.info("Configured instruments: {}".format(list(list_instruments.items())))
 
     # Non-amps modules
     non_amp_module = {"MODULE01": False, "MODULE02": False}
@@ -63,22 +53,27 @@ class Keithley2700VISADriver:
                                            read_termination="\n",
                                            )
             self._instr.timeout = 10000
+
             # Check if the selected resource match the loaded configuration
             model = self.get_idn()[32:36]
+            try:
+                assert model == self.model, model
+            except AssertionError as err:
+                logger.error("{}: Keithley {} used doesn't match the model in the configuration.".format(KeyError, err))
             if "27" not in model:
-                logger.warning("Driver designed to use Keithley 2700, not {} model. Problems may occur.".format(model))
-            for instr in config["Keithley", "2700"]:
-                if type(config["Keithley", "2700", instr]) == dict:
-                    if self.rsrc_name in config["Keithley", "2700", instr, "rsrc_name"]:
+                logger.warning("Driver designed to use Keithley 27XX series, not {} model.".format(model))
+            for instr in config["Keithley", self.model]:
+                if type(config["Keithley", self.model, instr]) == dict:
+                    if self.rsrc_name in config["Keithley", self.model, instr, "rsrc_name"]:
                         self.instr = instr
-            logger.info("Instrument selected: {} ".format(config["Keithley", "2700", self.instr, "rsrc_name"]))
-            logger.info("Keithley model : {}".format(config["Keithley", "2700", self.instr, "model_name"]))
+            logger.info("Instrument selected: {} ".format(config["Keithley", self.model, self.instr, "rsrc_name"]))
+            logger.info("Keithley model : {}".format(config["Keithley", self.model, self.instr, "model_name"]))
             try:
                 # Load the configuration matching the selected module
                 cards = self.get_card().split(',')
                 logger.info("card : {}".format(cards))
                 try:
-                    assert config["Keithley", "2700", self.instr, "MODULE01", "module_name"] == cards[0], cards[0]
+                    assert config["Keithley", self.model, self.instr, "MODULE01", "module_name"] == cards[0], cards[0]
                     self.configured_modules["MODULE01"] = cards[0]
                 except KeyError as err:
                     logger.error("{}: configuration {} does not exist.".format(KeyError, err))
@@ -86,7 +81,7 @@ class Keithley2700VISADriver:
                     logger.error("{}: Switching module {} does not match any configuration".format(
                         AssertionError, str(err)))
                 try:
-                    assert config["Keithley", "2700", self.instr, "MODULE02", "module_name"] == cards[1], cards[1]
+                    assert config["Keithley", self.model, self.instr, "MODULE02", "module_name"] == cards[1], cards[1]
                     self.configured_modules["MODULE02"] = cards[1]
                 except KeyError as err:
                     logger.error("{}: configuration {} does not exist." .format(KeyError, err))
@@ -95,10 +90,10 @@ class Keithley2700VISADriver:
                         AssertionError, str(err)))
                 logger.info("Configured modules : {}".format(self.configured_modules))
                 try:
-                    if config["Keithley", "2700", self.instr, 'MODULE01', 'module_name']\
+                    if config["Keithley", self.model, self.instr, 'MODULE01', 'module_name']\
                             in self.non_amp_modules_list:
                         self.non_amp_module["MODULE01"] = True
-                    if config["Keithley", "2700", self.instr, 'MODULE02', 'module_name']\
+                    if config["Keithley", self.model, self.instr, 'MODULE02', 'module_name']\
                             in self.non_amp_modules_list:
                         self.non_amp_module["MODULE02"] = True
                 except KeyError:
@@ -126,24 +121,24 @@ class Keithley2700VISADriver:
 
         # The following loop set up each channel in the config file
         for module in self.configured_modules:
-            for key in config["Keithley", "2700", self.instr, module, "CHANNELS"].keys():
+            for key in config["Keithley", self.model, self.instr, module, "CHANNELS"].keys():
 
                 # Handling user mistakes if the channels' configuration section is not correctly set up
-                if not type(config["Keithley", "2700", self.instr, module, 'CHANNELS', key]) == dict:
+                if not type(config["Keithley", self.model, self.instr, module, 'CHANNELS', key]) == dict:
                     logger.info("Channel {} not correctly defined, must be a dictionary" .format(key))
                     continue
-                if not config["Keithley", "2700", self.instr, module, 'CHANNELS', key]:
+                if not config["Keithley", self.model, self.instr, module, 'CHANNELS', key]:
                     continue
-                if "mode" not in config["Keithley", "2700", self.instr, module, 'CHANNELS', key]:
+                if "mode" not in config["Keithley", self.model, self.instr, module, 'CHANNELS', key]:
                     logger.info("Channel {} not fully defined, 'mode' is missing" .format(key))
                     continue
-                if config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "mode"].upper()\
+                if config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "mode"].upper()\
                         not in self.modes_channels_dict.keys():
                     logger.info("Channel {} not correctly defined, mode not recognized" .format(key))
                     continue
 
                 # Channel mode
-                mode = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "mode"].upper()
+                mode = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "mode"].upper()
                 self.modes_channels_dict[mode].append(int(key))
                 channel = '(@' + key + ')'
                 channels += key + ","
@@ -151,37 +146,37 @@ class Keithley2700VISADriver:
                 self._instr.write(cmd)
 
                 # Config
-                if 'range' in config["Keithley", "2700", self.instr, module, 'CHANNELS', key].keys():
-                    rang = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "range"]
+                if 'range' in config["Keithley", self.model, self.instr, module, 'CHANNELS', key].keys():
+                    rang = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "range"]
                     if 'autorange' in str(rang):
                         self._instr.write(mode + ':RANG:AUTO ')
                     else:
                         self._instr.write(mode + ':RANG ' + str(range))
 
-                if 'resolution' in config["Keithley", "2700", self.instr, module, 'CHANNELS', key].keys():
-                    resolution = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "resolution"]
+                if 'resolution' in config["Keithley", self.model, self.instr, module, 'CHANNELS', key].keys():
+                    resolution = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "resolution"]
                     self._instr.write(mode + ':DIG ' + str(resolution))
 
-                if 'nplc' in config["Keithley", "2700", self.instr, module, 'CHANNELS', key].keys():
-                    nplc = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "nplc"]
+                if 'nplc' in config["Keithley", self.model, self.instr, module, 'CHANNELS', key].keys():
+                    nplc = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "nplc"]
                     self._instr.write(mode + ':NPLC ' + str(nplc))
 
                 if "TEMP" in mode:
-                    transducer = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "transducer"].upper()
+                    transducer = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "transducer"].upper()
                     if "TC" in transducer:
-                        tc_type = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "type"].upper()
-                        ref_junc = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "ref_junc"].upper()
+                        tc_type = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "type"].upper()
+                        ref_junc = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "ref_junc"].upper()
                         self.mode_temp_tc(channel, transducer, tc_type, ref_junc)
                     elif "THER" in transducer:
-                        ther_type = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "type"].upper()
+                        ther_type = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "type"].upper()
                         self.mode_temp_ther(channel, transducer, ther_type)
                     elif "FRTD" in transducer:
-                        frtd_type = config["Keithley", "2700", self.instr, module, 'CHANNELS', key, "type"].upper()
+                        frtd_type = config["Keithley", self.model, self.instr, module, 'CHANNELS', key, "type"].upper()
                         self.mode_temp_frtd(channel, transducer, frtd_type)
 
                 # Console info
                 logger.info("Channels {} \n {}".format(key,
-                                                       config["Keithley", "2700", self.instr, module, 'CHANNELS', key]))
+                                                       config["Keithley", self.model, self.instr, module, 'CHANNELS', key]))
                 # Timeout update for long measurement modes such as voltage AC
                 if "AC" in mode:
                     self._instr.timeout += 4000
@@ -232,9 +227,9 @@ class Keithley2700VISADriver:
             # Trigger scan
             self._instr.write("*TRG")
             # Get data (equivalent to TRAC:DATA? from buffer)
-            str_answer = self._instr.query("FETCH?")[1:-1]
+            str_answer = self.get_data()
         else:
-            str_answer = self._instr.query("FETCH?")[1:-1]
+            str_answer = self.get_data()
         # Split the instrument answer (MEASUREMENT,TIME,READING COUNT) to create a list
         list_split_answer = str_answer.split(",")
 
@@ -278,18 +273,6 @@ class Keithley2700VISADriver:
 
         return str_answer, array_measurements_values, array_times_values
 
-    def get_card(self):
-        # Query switching module
-        return self._instr.query("*OPT?")[1:-1]
-    
-    def get_error(self):
-        # Ask the keithley to return the last current error
-        return self._instr.query("SYST:ERR?")[1:-1]
-    
-    def get_idn(self):
-        # Query identification
-        return self._instr.query("*IDN?")[1:-1]
-    
     def init_cont_off(self):
         # Disable continuous initiation
         self._instr.write("INIT:CONT OFF")
@@ -395,63 +378,3 @@ class Keithley2700VISADriver:
                     self._instr.write("ROUT:SCAN:LSEL INT")
                 
             return channels
-        
-    def stop_acquisition(self):
-        # If scan in process, stop it
-        self._instr.write("ROUT:SCAN:LSEL NONE")
-
-    def user_command(self):
-        command = input('Enter here a command you want to send directly to the Keithley [if None, press enter]: ')
-        if command != '':
-            if command[-1] == "?":
-                print(self._instr.query(command)[1:-1])
-            else:
-                self._instr.write(command)
-            self.user_command()
-        else:
-            pass
-
-
-if __name__ == "__main__":
-    try:
-        print("In main")
-
-        # You can use this main section for:
-        # - Testing connexion and communication with your instrument
-        # - Testing new methods in developer mode
-
-        RM = visa.ResourceManager("@py")
-        print("list resources", RM.list_resources())
-
-        # K2700 Instance of KeithleyVISADriver class (replace ASRL1::INSTR by the name of your resource)
-        k2700 = Keithley2700VISADriver("ASRL1::INSTR")
-        k2700.init_hardware()
-        print("IDN?")
-        print(k2700.get_idn())
-        k2700.reset()
-        k2700.configuration_sequence()
-
-        # Daq_viewer simulation first run
-        k2700.set_mode(str(input('Enter which mode you want to scan \
-        [scan_scan_list, scan_volt:dc, scan_r2w, scan_temp...]:')))
-        print('Manual scan example of command set to send directly: >init >*trg >trac:data?')
-        k2700.user_command()
-        print('Automatic scan example with 2 iterations')
-        for i in range(2):
-            print(k2700.data())
-        print(k2700.data())
-
-        # Daq_viewer simulation change mode
-        k2700.set_mode(str(input('Enter which mode you want to scan \
-        [scan_scan_list, scan_volt:dc, scan_r2w, scan_temp...]:')))
-        for i in range(2):
-            print(k2700.data())
-        print(k2700.data())
-
-        k2700.clear_buffer()
-        k2700.close()
-
-        print("Out")
-
-    except Exception as e:
-        print("Exception ({}): {}".format(type(e), str(e)))
